Set SPI1 FIFO threshold and data size on init
The SPI peripheral on the STM32F0x2 has a configurable FIFO threshold
(it can be either 8 or 16 bit, with the FIFO capacity being 32 bits).
This threshold is controlled by the `FRXTH` bit in the `CR2` register
(reference manual, pp 804).

When using the default blocking implementation, the driver will write
and the read bytes alternately. When the FIFO threshold is too high, the
`RXNE` bit is not set after the first word -- and the driver blocks
indefinitely.

The presence of the FIFO suggests the possibility of a more efficient,
non-default blocking implementation, which uses the FIFO to help pack
transfers together more tightly and hence achieve a higher throughput.

Fixes 98227f23ccea ("Added SPI code and an APA102C example")
Signed-off-by: Austin Glaser <austin.glaser@gmail.com>